add auth and comment documentation

+++ src/controller/faq-controller.js
@@ -1,7 +1,8 @@
 import Faq from '../model/faq.js'
 
 const createFaq = async (req, res, next) => {
-    const { pertanyaan,
+    const {
+        pertanyaan,
         jawaban } = req.body
     try {
         const newFaq = await Faq.create({
@@ -45,6 +46,7 @@ const getFaqById = async (req, res, next) => {
     const { id } = req.params
     try {
         const faq = await Faq.findOne({ _id: id })
+        //when id faq is not found
         if (!faq) {
             return res.status(400).json({
                 status: 400,
@@ -69,10 +71,21 @@ const getFaqById = async (req, res, next) => {
 
 const editFaq = async (req, res, next) => {
     const { id } = req.params
-    const { pertanyaan,
+    const {
+        pertanyaan,
         jawaban } = req.body
     try {
-        const updatedFaq = await Faq.updateOne({ _id: id }, {
+        const faq = await Faq.findOne({ _id: id })
+        //when id faq is not found
+        if (!faq) {
+            return res.status(400).json({
+                status: 400,
+                message: 'failed',
+                info: 'faq not found'
+            });
+        }
+
+        await Faq.updateOne({ _id: id }, {
             $set: {
                 pertanyaan,
                 jawaban
@@ -96,7 +109,18 @@ const editFaq = async (req, res, next) => {
 const deleteFaq = async (req, res, next) => {
     const { id } = req.params
     try {
+        const faq = await Faq.findOne({ _id: id })
+        //when id faq is not found
+        if (!faq) {
+            return res.status(400).json({
+                status: 400,
+                message: 'failed',
+                info: 'faq not found'
+            });
+        }
+
         const deletedFaq = await Faq.deleteOne({ _id: id })
+        //when no one faq is deleted
         if (deletedFaq.deletedCount === 0) {
             return res.status(400).json({
                 status: 400,
